requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
openai>=1.58.1,<2.0.0
pillow==10.1.0
imageio==2.31.5
pydantic==2.5.2
pytest==7.4.3
python-dotenv==1.0.0
langchain>=0.0.340
langchain-openai==0.3.1
browser-use
playwright